remove tab window as it is unnecessary
