+++ new-gui/.expo/packager-info.json
@@ -1,9 +1,9 @@
 {
   "expoServerNgrokUrl": "https://rp-cnx.anonymous.new-gui.exp.direct",
   "packagerNgrokUrl": "https://packager.rp-cnx.anonymous.new-gui.exp.direct",
-  "ngrokPid": 19000,
+  "ngrokPid": 20080,
   "packagerPort": 19001,
-  "packagerPid": 17996,
+  "packagerPid": 10632,
   "devToolsPort": 19002,
   "expoServerPort": 19000
 }
